Added my methods to shopping class
